@@ -3,6 +3,7 @@ plugins {
     id "org.jetbrains.kotlin.kapt" version "$kotlinVersion"
     id "org.jetbrains.kotlin.plugin.allopen" version "$kotlinVersion"
     id "com.github.johnrengelman.shadow" version "$shadowVersion"
+    id "com.github.ben-manes.versions" version "$gradleVersionsPlugin"
     id "application"
 }
 
@@ -22,8 +23,6 @@ configurations {
 
 dependencies {
     implementation platform("io.micronaut:micronaut-bom:$micronautVersion")
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${kotlinVersion}"
-    implementation "org.jetbrains.kotlin:kotlin-reflect:${kotlinVersion}"
     implementation "io.micronaut:micronaut-runtime"
     implementation "io.micronaut:micronaut-http-server-netty"
     implementation "io.micronaut:micronaut-http-client"
@@ -54,7 +53,7 @@ dependencies {
 
     implementation "io.micronaut.data:micronaut-data-jdbc:$micronautDataVersion"
     implementation "io.micronaut.sql:micronaut-jdbc-hikari:$micronautSQLVersion"
-    implementation "jakarta.persistence:jakarta.persistence-api:2.2.2"
+    implementation "jakarta.persistence:jakarta.persistence-api:$jakartaPersistence"
     implementation "io.micronaut.flyway:micronaut-flyway"
 
     implementation "no.nav.arbeid.pam:pam-ansettelseskodeverk:$pamAnsettelsesKodeverkVersion"
@@ -62,7 +61,7 @@ dependencies {
     implementation "org.postgresql:postgresql:$postgresqlVersion"
     implementation "io.micronaut.security:micronaut-security-jwt"
     implementation "io.swagger.core.v3:swagger-annotations"
-    implementation "com.googlecode.owasp-java-html-sanitizer:owasp-java-html-sanitizer:20190610.1"
+    implementation "com.googlecode.owasp-java-html-sanitizer:owasp-java-html-sanitizer:$htmlSanitizer"
     implementation('org.apache.commons:commons-text:1.10.0')
     implementation("io.micronaut.cache:micronaut-cache-caffeine")
     runtimeOnly project(':frontend')
@@ -71,11 +70,10 @@ dependencies {
     testImplementation("org.testcontainers:postgresql:${tcVersion}")
  }
 
-
 test.classpath += configurations.developmentOnly
 
 application {
-    mainClassName = "no.nav.arbeidsplassen.importapi.Application"
+    mainClass = "no.nav.arbeidsplassen.importapi.Application"
 }
 
 test {
@@ -96,7 +94,7 @@ allOpen {
 
 compileKotlin {
     kotlinOptions {
-        jvmTarget = "14"
+        jvmTarget = "17"
         //Will retain parameter names for Java reflection
         javaParameters = true
     }
@@ -104,7 +102,7 @@ compileKotlin {
 
 compileTestKotlin {
     kotlinOptions {
-        jvmTarget = "14"
+        jvmTarget = "17"
         javaParameters = true
     }
 }
